requirements.txt
fastmcp>=0.2.0
starlette>=0.27.0
uvicorn>=0.23.0
gunicorn>=21.2.0